Test : Workaround bogus Linux returncode on default systems
When TF_FATAL diagnostics are emitted, OpenUSD attempts to terminte the current process using its Arch module.

On Linux, the logic attempts to fork/ptrace itself to see if there's a debugger attached. If there is, it will abort with exit code 0. If no debugger is attached it will abort with a proper non-zero code.

Per the ptrace manpage the default ptrace_scope on most modern Linux systems is to have "restricted permissions" for security reasons, and this confuses the Arch logic into thinking a debugger is attached when it is not.

Since we want to excercise the `IssueFatalError` implementation in our `Delegate`, we need to ignore the Linux return code unless "classic ptrace permissions" are enabled in the kernel.