PrimvarData Tests : Update test for behaviour on modern Linux
When we moved from the old cxx abi GCC 7 builds to the new abi GCC 11 builds, there is a behavior change in moving an optional.

We have known about the issue for a long time, we are just now needing to adjust for it. It is possible some changes to PrimvarData could avoid this, but this is an edge case that does not seem like chasing it down will pay off. The VtArrays have still not detached (copied) so there is no real cost here.