Build : Add a standalone python wheel
This uses poetry to build a .whl which can be installed in an isolated virtual environment.

```
python -m venv foo
source foo/bin/activate
python -m pip install usd_exchange[usd2502]
python -c "import usdex.core; from pxr import Usd"
```

The whl is specific to a particular USD flavor/version and must be used with the same USD binaires that we compiled against.

We do not use the public `usd-core` wheel from Pixar as a dependency, since it is not ABI compatible with our minspec (CXX ABI is the primary issue). It also does not contain the exact set of libs we wish to distribute (e.g. we would like to add usdMtlx). Instead, we bundle the exact OpenUSD libs that we compiled aginst into our whl directly. In the future, if the public `usd-core` package becomes more compatible, we could add it as a dependency instead.